+++ client/mobile/App/navigation/index.tsx
@@ -5,43 +5,19 @@
  * @flow
  */
 
-import React, { Component } from "react";
-import { Platform, StyleSheet, Text, View } from "react-native";
+import * as React from "react";
+import { createStackNavigator } from "react-navigation";
 
-const instructions = Platform.select({
-  android:
-    "Double tap R on your keyboard to reload\n" +
-    "Shake or press menu button for dev menu",
-  ios: "Press Cmd+R to reload,\n" + "Cmd+D or shake for dev menu",
+import { HomeScreen } from "App/view";
+
+const RootStack = createStackNavigator({
+  Home: {
+    screen: HomeScreen,
+  },
 });
 
-export default class App extends Component<{}> {
+export class App extends React.Component<{}> {
   public render() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>Welcome to React Native!</Text>
-        <Text style={styles.instructions}>To get started, edit App.js</Text>
-        <Text style={styles.instructions}>{instructions}</Text>
-      </View>
-    );
+    return <RootStack />;
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#F5FCFF",
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: "center",
-    margin: 10,
-  },
-  instructions: {
-    textAlign: "center",
-    color: "#333333",
-    marginBottom: 5,
-  },
-});
