+++ client/mobile/index.js
@@ -1,7 +1,7 @@
 /** @format */
 
 import { AppRegistry } from "react-native";
-import App from 'App/navigation';
+import { App }  from 'App/navigation';
 import { name as appName } from "GooseOrNot/app.json";
 
 AppRegistry.registerComponent(appName, () => App);
